requirements.txt
faster-whisper 
pydub 
webrtcvad
pysrt 
moviepy 
gTTs
TTS 
edge_tts
ffmpeg-python 
pysubs2
pyannote.audio
demucs 
librosa
speechbrain
soundfile 
scipy 
yt-dlp
python-dotenv
datasets
langchain 
langchain-core 
langchain-openai 
langchain_community
pydantic
PySide6 
python-vlc
python-decouple
transformers>=4.36.0
accelerate>=0.27.0